requirements.txt
psycopg2-binary
requests
flask
pandas
scikit-learn
python-dotenv
numpy
joblib
sqlalchemy
asyncpg
Flask[async]
quart
pytest
httpx
pytest-asyncio